requirements.txt
datasets
huggingface
openai
anthropic
google-generativeai
together